Changed Data Types

+++ StarsAndStripes/Program.cs
@@ -6,12 +6,12 @@
         private const ConsoleColor Blue = ConsoleColor.Blue;
         private const ConsoleColor White = ConsoleColor.White;
         private const ConsoleColor Black = ConsoleColor.Black;
-        private static int[] Reds = new int[] { 0, 1, 4, 5, 8, 9, 12, 13, 16, 17, 20, 21, 24, 25 };
+        private static byte[] Reds = new byte[] { 0, 1, 4, 5, 8, 9, 12, 13, 16, 17, 20, 21, 24, 25 };
 
         static void Main(string[] args)
         {
             //Create the first 11 rows.
-            for (int i = 0; i < 11; i++)
+            for (byte i = 0; i < 11; i++)
             {
                 BlueStripe(23, i);
                 if (Reds.Contains(i))
@@ -25,7 +25,7 @@
                 Console.WriteLine();
             }
             //create the rest of the flag.
-            for (int i = 11; i < 26; i++)
+            for (byte i = 11; i < 26; i++)
             {
                 if (Reds.Contains(i))
                 {
@@ -44,18 +44,17 @@
         /// Create white stripes with a lenght of fields.
         /// </summary>
         /// <param name="fields"></param>
-        static void WhiteStripe(int fields)
+        static void WhiteStripe(byte fields)
         {
             Console.BackgroundColor = White;
             WriteEmpty(fields);
-
         }
         
         /// <summary>
         /// Create Red Stripes with a length of fields.
         /// </summary>
         /// <param name="fields"></param>
-        static void RedStripe(int fields)
+        static void RedStripe(byte fields)
         {
             Console.BackgroundColor = Red;
             WriteEmpty(fields);
@@ -67,7 +66,7 @@
         /// </summary>
         /// <param name="fields"></param>
         /// <param name="row"></param>
-        static void BlueStripe(int fields, int row)
+        static void BlueStripe(byte fields, byte row)
         {
             Console.BackgroundColor = Blue;
             if (row == 0 || row == 10) { WriteEmpty(fields); }
@@ -85,11 +84,11 @@
         /// Create even rows.
         /// </summary>
         /// <param name="fields"></param>
-        static void EvenStar(int fields)
+        static void EvenStar(byte fields)
         {
-            int[] stars = new int[] { 1, 5, 9, 13, 17, 21 };
+            byte[] stars = new byte[] { 1, 5, 9, 13, 17, 21 };
             Console.BackgroundColor = Blue;
-            for (int i = 0; i < fields; i++)
+            for (byte i = 0; i < fields; i++)
             {
                 if (i == 0)
                 {
@@ -110,11 +109,11 @@
         /// Create uneven rows.
         /// </summary>
         /// <param name="fields"></param>
-        static void UnEvenStar(int fields)
+        static void UnEvenStar(byte fields)
         {
-            int[] stars = new int[] { 3, 7, 11, 15, 19, 23 };
+            byte[] stars = new byte[] { 3, 7, 11, 15, 19, 23 };
             Console.BackgroundColor = Blue;
-            for (int i = 0; i < fields; i++)
+            for (byte i = 0; i < fields; i++)
             {
                 if (i == 0)
                 {
@@ -135,9 +134,9 @@
         /// Write empty rows.
         /// </summary>
         /// <param name="fields"></param>
-        private static void WriteEmpty(int fields)
+        private static void WriteEmpty(byte fields)
         {
-            for (int i = 0; i < fields; i++)
+            for (byte i = 0; i < fields; i++)
             {
                 Console.Write("   ");
             }
